fix: change log service to a TS file
